+++ attributes/centos-nfs-ganesha.rb
@@ -8,7 +8,7 @@ default['yum']['centos-nfs-ganesha']['enabled'] = false
 default['yum']['centos-nfs-ganesha']['make_cache'] = true
 default['yum']['centos-nfs-ganesha']['managed'] = false
 default['yum']['centos-nfs-ganesha']['mirrorlist'] =
-  "http://mirrorlist.centos.org/?release=$releasever&arch=$basearch&repo=storage-nfsganesha-#{ver}"
+  "http://mirrorlist.centos.org/?release=#{release_var}&arch=$basearch&repo=storage-nfsganesha-#{ver}"
 default['yum']['centos-nfs-ganesha']['gpgkey'] = 'https://www.centos.org/keys/RPM-GPG-KEY-CentOS-SIG-Storage'
 # testing
 default['yum']['centos-nfs-ganesha-testing']['repositoryid'] = 'centos-nfs-ganesha-testing'
@@ -19,4 +19,4 @@ default['yum']['centos-nfs-ganesha-testing']['make_cache'] = true
 default['yum']['centos-nfs-ganesha-testing']['managed'] = false
 default['yum']['centos-nfs-ganesha-testing']['gpgcheck'] = false
 default['yum']['centos-nfs-ganesha-testing']['baseurl'] =
-  "https://buildlogs.centos.org/centos/$releasever/storage/$basearch/nfsganesha-#{ver}/"
+  "https://buildlogs.centos.org/centos/#{release_var}/storage/$basearch/nfsganesha-#{ver}/"