+++ attributes/powertools.rb
@@ -1,7 +1,8 @@
 default['yum']['powertools']['repositoryid'] = 'powertools'
-default['yum']['powertools']['description'] = "CentOS-#{node['platform_version'].to_i} - PowerTools"
+default['yum']['powertools']['description'] = "#{centos_description} - PowerTools"
 default['yum']['powertools']['enabled'] = false
 default['yum']['powertools']['make_cache'] = true
 default['yum']['powertools']['managed'] = false
-default['yum']['powertools']['mirrorlist'] = "http://mirrorlist.centos.org/?release=#{node['platform_version'].to_i}&arch=$basearch&repo=PowerTools"
+default['yum']['powertools']['mirrorlist'] =
+  "http://mirrorlist.centos.org/?release=#{release_var}&arch=$basearch&repo=PowerTools"
 default['yum']['powertools']['gpgkey'] = 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial'